requirements.txt
streamlit
moviepy
gTTS
Pillow
customtkinter